add build artifacts for chrome & firefox
